Correction type du champ delai_off en base de données Temporisation
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@3013 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/init_db.sql
@@ -804,7 +804,7 @@ CREATE TABLE IF NOT EXISTS `rs485` (
 CREATE TABLE IF NOT EXISTS `mnemos_Tempo` (
   `id_mnemo` int(11) NOT NULL,
   `delai_on` int(11) NOT NULL DEFAULT '0',
-  `delai_off` tinyint(1) NOT NULL DEFAULT '0',
+  `delai_off` int(11) NOT NULL DEFAULT '0',
   `min_on` int(11) NOT NULL DEFAULT '0',
   `max_on` int(11) NOT NULL DEFAULT '0',
   PRIMARY KEY (`id_mnemo`),
